Create README.md
Power BI Analysis Reveals Top Carbon Emitters

Our analysis of the data in Power BI reveals that the top emitting entities are all firms involved in coal and natural gas extraction. Additionally, coal is a far greater carbon emitter than the oil and gas industry.

The final graph shows a consistent increase in emissions each year, with the exception of a brief period following 1991. This pattern can be correlated with the fall of the Soviet Union, the world's second-biggest carbon emitter at the time.  Following the collapse of the Soviet Union, their industries also suffered significantly, leading to a decrease in their carbon emissions.